requirements.txt
aiogram>=3.4,<4.0
aiohttp>=3.9,<4.0
openai>=1.30,<2.0
python-dotenv>=1.0,<2.0
